requirements.txt
argparse
logging
rich
pyyaml
questionary
pytest
ffmpeg-python